Updated releng/autorelease Project: yangtools  8ee025ef04356036d18d32a27e4163e270e22bb6
Cap LazyCollections list

When we move from a singleton to an ArrayList, specify the size as 2, as
the arraylist can grow beyond that, but will not waste space if it
remains at two elements.

Change-Id: Ie600f25f9e59aa5321f335b380d6d6930e4492ee
Signed-off-by: Robert Varga <rovarga@cisco.com>

Introduce UnmodifiableCollection

This is a utility class which recognizes itself and Guava's
ImmutableCollection subclasses during contruction, preventing unneeded
encapsulation.

Change-Id: I99deeb570a74ecee5f0bb8a89c25b363284e5f6e
Signed-off-by: Robert Varga <rovarga@cisco.com>

Expand known immutables

Void and Boolean are immutable types, too. Also use ImmutableSet instead
of a hown-grown utility.

Change-Id: I63bddcfe5a8538065378a2259bf8bdc58f76d620
Signed-off-by: Robert Varga <rovarga@cisco.com>