Updated releng/autorelease Project: ovsdb  6abc3c1c44411a59c6d6931893ca6757b88d49e0
OVSDB V3 northbound failed to find node

The Node.fromString(id) no longer works with the latest adsal deprecation changes.

A new helper method getNode() was previously created as a workaround but the v3 code was not using it. This commit fixes that.

Also added unit test for the v3 northbound.

Change-Id: I3bdffe52b8ba63b6bd52da482b0a3b1a162b6db8
Signed-off-by: Sam Hague <shague@redhat.com>