Updated releng/autorelease Project: bgpcep  9561bf2086de1e80b6fe9ea9115ec84641320e57
BUG-2383: fix the RIB model

The overall goal is to store adjacency ribs in the datastore. In order
to do that, we need to fix the (defunct) model, so each peer has a
unique subtree. This patch fixes the model and introduces a peer ID
type, which we will use throughout to identify peers.

Change-Id: I45a4493d0aabb29828e6e935017939a5b09e7b60
Signed-off-by: Robert Varga <rovarga@cisco.com>