Updated releng/autorelease Project: bgpcep  2bc21d065337b6e27852f54a48553fb9155bbfd3
BUG-2383: rework RIB inbound

Each Peer (like BGPPeer) performs updates the adj-rib-in via a dedicated
AdjRibInWriter instance. That class, along with RIBSupport and TableContext,
is responsible for translating the Binding-Aware messages coming from
the message parser into Binding-Independent form within the datastore.

Changes are then propagated via DataChangeNotifications into
InboundPolicyEnforcer, which performs any munging of attributes required
and checks if the route is eligible for inclusion in the local RIB. If
it is, the route is output into the effective-rib-in hosted in the
datastore.

From there it is picked up by BestPathSelection, which compares the
effective routes from multiple peers and installs the best one into the
local RIB.

Change-Id: Ie23746b2fc44499f42d346595562b498c19fd112
Signed-off-by: Robert Varga <rovarga@cisco.com>