Updated releng/autorelease Project: controller  ceb41a4fe9c36a2036c5dcae0fc02359be8bdd66
BUG 2412 - restconf getOperations method migration

* migration to new faster Infrastructure API and Codecs for method
@GET getOperations(UriInfo) on @Path {/operations}

New faster Infrastructure API works with NormizedNodeContext and
we are replacing first method getModule(URI) from RestconfService
to use NormalizedNodeContext.
BUT we are not able follow specification for migration in this case
because, we have to change restconf-netconf yang schema before.

Change-Id: If17aefbaa139ee7f1064b25a8592da76d71c4bc4
Signed-off-by: Vaclav Demcak <vdemcak@cisco.com>

BUG 2412 - restconf mountpoint getOperations method migration

* migration to new faster Infrastructure API and Codecs for method
@GET getOperations(String,UriInfo) on @Path {/operations/identifier}

New faster Infrastructure API works with NormizedNodeContext and
we are replacing first method getModule(URI) from RestconfService
to use NormalizedNodeContext.
BUT we are not able follow specification for migration in this case
because, we have to change restconf-netconf yang schema before.

Change-Id: I36d0f5ec88af7c4a9d8b7a2d7ed84ccdbabd9c13
Signed-off-by: Vaclav Demcak <vdemcak@cisco.com>