Updated releng/autorelease Project: yangtools  78859c4331774ebf2e6c5d0b05b3e3c22e6b5937
Bug 2818: Introduced optional array for parsing root list items.

Specification draft-lhotka-netmod-yang-json-02 clearly
specifies that list must be serialized as array of objects,
but currently only user of yang-json is
draft-bierman-netconf-restconf-02, which specifies
and provides example where such list exists are also used
without arrays when parsing.

Introduced support for this case (and unit test),
in order to interoperate with clients, which follows
serialization as was provided in examples of restconf.

Change-Id: I4ebf7cfba17752de2184e9b7ac1d18c1ecac6f80
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>