Updated releng/autorelease Project: sfc  a555b4a811e906803e107813e11d3ca45713c071
Temp fix for unit test cases fails

The cases of
SfcProviderServiceChainAPITest.testCreateReadServiceFunctionChain
and SfcServiceFuntionSchedulerAPITest.testBasicEnvSetup may fail
randomly.
The logs are like:
SfcProviderServiceChainAPITest.testCreateReadServiceFunctionChain:167
 Must be equal expected:<[SfcServiceFunction{getName=simple_fw_101,
or java.lang.AssertionError: Must be equal expected:
<[SfcServiceFunction{getName=simple_fw_101, getType=class

In these cases, there is no such Service Function Types which are
named simple_fw_10[1|2|3]. These types are stored in the
SfcProviderServicePathAPITest and the delete transaction may fail
sometimes. If the transaction fails, the chain which named
"unittest-chain-1" is left in the datastore and will append new
Service Function Types in it. Then the chain does no longer look
like what is expected.

This patch renames the chains' names in different tests.

Signed-off-by: Johnson Li <johnson.li@intel.com>