Updated releng/autorelease Project: yangtools  590913c53d782033f49b96347aeea994efdee49b
Reduce InMemoryDataTree logging

In setSchemaContext it logs the "Attempting to install schema contexts"
message to info which causes significant clutter in the karaf log - as
many as thousands of messages on startup and shutdown. I don't think
this message provides any useful information anyway so I removed it as
a similar message is logged to debug with context.

Change-Id: I0df9c3f49bd8173318b7b68584d066739ae7a562
Signed-off-by: tpantelis <tpanteli@brocade.com>