Updated releng/autorelease Project: dlux  33a19fbd8d03d9318dfd66957ec6a0d192fdef26
Initial refactor for the topology renderer

This commit contain a "home made" graph renderer. It's build with the
combination of NGraph as graph manager and Pixijs as WebGL (2D) renderer.
The goal is to replace Visjs to something more customizable and later on
add some 3D features. For now, the graph does the basic thing
(move nodes, pan and zoom).

Change-Id: Iee9ee2cb59ffb7a8ef812b0ee9def4ffd4164556
Signed-off-by: Maxime Millette-Coulombe <mmcoulombe@inocybe.com>