Updated releng/autorelease Project: yangtools  e1a3876862c946f484edd03c0012575037a32d03
Bug 2361: Added prototype implementation using new SPI.

This prototype level implementation of module, namespace,
import, uses and container statement shows usage
of meta-level SPI in order to introduce support for
concrete statements.

Note that this is same SPI, which is to be used
to introduce new statement or extension support
directly to the YANG parser.

Change-Id: I5124a06669d48d3a32d585c4c75de5b2228f2eac
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>
Signed-off-by: Martin Ciglan <mciglan@cisco.com>