Updated releng/autorelease Project: yangtools  b9eea0e9b4d9c0745d1f6462526a636eb4fa91e9
BUG-2856: add leaf node simplied by key

Since we have the cross-reference between key and object leaf nodes, we
enforce the consistency. Unfortunately this is not something obvious to
users. We have enough information to correct the situation, so lower the
check to apply only when an explicit inconsistency is detected.

Change-Id: I9297a7d60e4e8d4cd0680a74c1e45f412f8b5f77
Signed-off-by: Robert Varga <rovarga@cisco.com>