Updated releng/autorelease Project: yangtools  8c7f61fe6c73a27c748814d2610d7784ce73e796
BUG-2876: determine ModificationType while applying changes

SchemaAwareApplyOperation is the class which actually determines what is
happening to the data, so it makes sense for it to mark what was
actually done, so that other classes do not get to guess (and arrive at
wrong results).

Also rename ModifiedNode#updateModificationType() to a new name so we
reduce overall naming confusion.

Change-Id: I6436636ed83d2d4a84afc8c9e5e50def1b1ccab0
Signed-off-by: Robert Varga <rovarga@cisco.com>