Updated releng/autorelease Project: yangtools  9b26bdbf76e7792976dbd8c032449c2e664f170b
Do not use recursive tree creation

AbstractValueNodeModificationStrategy is creating leaf entries in the
tree. We know for a fact they do not have children, so there is no need
to eagerly instantiate child tree nodes.

The same thing is true for the root node. Since we can deal with lazily
instantiating tree nodes, we do not have to eagerly allocate them.

Change-Id: I1a55cc9d5f43c8b2c92324d470ab533fc06ad95c
Signed-off-by: Robert Varga <rovarga@cisco.com>

Fix child ordering assumptions

When instantiating nodes during an operation, we first looked up the
strategy and then specified its child handling strategy to all
intermediate nodes. This means we could have ended up specifying
ordering on a node which does not needed, or (worse) not specifying it
where it was needed.

Clean this up by walking the modification and strategy trees in sync,
bailing if we run out of strategy nodes.

While at it, turn the simple boolean into a full-blown enum, as what we
want to pass down is a tri-state: ORDERED, UNORDERED and NONE. The first
two correspond to previous true and false. The last one is useful for
leaves, as they cannot contain children, which means we can skip
instantiation of a mutable map and just use Collections.emptyMap().

Change-Id: I6ffd48b481a149255325dc31f99b7b23eebc928e
Signed-off-by: Robert Varga <rovarga@cisco.com>