Updated releng/autorelease Project: l2switch  b39f37219db6dada740ea8f10c7ba534be2ab184
Removed unused argument 'payload'

The argument 'payload' was used in an earlier version, as the
ReactiveFlowWriter handled ARP by its own and dispatched the
payload of the ARP requests.The ARP handling is now done by the
arphandler. The leftover argument 'payload' in the ReactiveFlowWriter
is usless and should be removed.

Change-Id: I53762d12d550d523de38e7dad5839c22b51dc591
Signed-off-by: Mario Willems <ma.willems@googlemail.com>