Updated releng/autorelease Project: integration  2f513e128646c6835b58269d67613944b9c38f4f
Testplan, suite, variables and support library for basic PCEP testing.

https://wiki.opendaylight.org/view/BGP_LS_PCEP:Lithium_Feature_Tests#PCEP
specifies odl-bgpcep-pcep-all as an user-facing feature.
By M4, an automated test (not necessarily complete)
should be running in CI.

This is the fourth version of suite for testing basic PCEP,
this version is finally ready for merge.

Changes (added and missing) against previous patch sets:
+ Use space-only separation, formatted by RIDE
- Still using .robot extension
+ Documentation and docstrings added everywhere
+ Setup steps compacted to one Suite Setup keyword
+ Suite Teardown added, various log gathering happens there
+ Non-critical tests now can fail on Read Until
+ Start_Pcc_Mock now does tee to a file, its contents goes to Log
+ HfsJson library added, for sorting and indenting JSON strings
+ ${MININET} substituted to place where PCC IP address should be
+ Base64 code for tunnel name is computed in the variable file
+ Long JSON data is in the variable file, in readable form
+ PEP 8 compliant
+ pylint output addressed (not all, pylint is wrong sometimes)
+ Python files now do contain license headers
- actually two sections, with slight incompatibilities between them
+ Robot file now contains a single license section, in suite documentation
- No tests for TCPMD5, they will be added as another Change

Example Sandbox run for this patch set:
https://jenkins.opendaylight.org/sandbox/job/bgpcep-csit-1node-imds-basicpcep-only-master/17/

After this change is merged, separate change will introduce
csit job to releng/builder, triggered by upstream distribution jobs.

After that is merged, another change will add this suite
to the list of suites to be run in integration-verify job.
Until then, integration-verify job is insensitive to this suite.


Change-Id: I24a8704c07780a2a97904284eb3f14de1dde50b3
Signed-off-by: Vratko Polak <vrpolak@cisco.com>