Updated releng/autorelease Project: openflowplugin  97e4f43e824a26765037a7673b93744019ee5e61
PostHandshakeNodeProducer help class

Note: PostHandshake process has to populate complex static information about
new FlowCapableNode before add DataStore process. But whole postHandsake
process contains a lot of functionality which is related only this short
time window so we put it to one class together. This new helper has to be
forgotten after succesfull add to DataStore process and it doesn't need
to be a part of DeviceManager code.

* settableFuture<Node> produce new Node with all device info and features
  * collecting Devide description
  * collecting Device meterFutures
  * collecting Device grupFutures
  * collecting Device tableFutures (TODO : add DataStore model and finish impl)

(Tests and possible fixies for this class will be
 a part of DeviceManager#deviceConnected method implemenation and its test suite)

Change-Id: I84fdd3004f211590d5066711518490cd134d6573
Signed-off-by: Vaclav Demcak <vdemcak@cisco.com>