Updated releng/autorelease Project: ovsdb  59e80a5ba8035c449b6f961cc5091fb1efb6abb5
Small fix to insure correct running offline.

The southbound-impl features depend on the config file, not
just on the southbound-impl bundle.

<groupId>org.opendaylight.ovsdb</groupId>
<artifiactId>southbound-impl</artifactId>

is the *bundle* dependency.

<groupId>org.opendaylight.ovsdb</groupId>
<artifiactId>southbound-impl</artifactId>
<type>xml</type>
<classifier>config</classifier>

is a distinct artifact for the config file.

By failing to depend on the config file, the 
southbound-impl feature will not work correctly in
offline mode.

Change-Id: Iab4ba7225f30aac19b45267da33c3a6219488b4a
Signed-off-by: Ed Warnicke <eaw@cisco.com>