Updated releng/autorelease Project: ovsdb  9f2354e97ed1d5f3e16152173a609b4a1fec9570
Add break in iteration over vlan tags for a given ovsdbTerminationPoint

Since we expect no more than a single vlan tag for a given termination point,
there is no need to call setVlanTag multiple times on termination point builder.
Added a TODO to mark the fact that the iteration/builder will need to be enhanced
if more than a single vlan tag support is needed.

Change-Id: I1ff45ccea8fb2b061157ca7b9af0ad2e072abdcb
Ref: https://gist.github.com/1e07ddc7e663710c2ac7
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>