Updated releng/autorelease Project: openflowplugin  f74dab9e40cc10fd0a3d6e77c1fece2e16db489c
Bug 2945: Fix read failures in StatAbstractListenCommit

Added a volatile currentReadTxStale flag which onDataChanged sets instead of
closing the currentReadTx. readLatestConfiguration checks the
currentReadTxStale flag and, if set, closes the currentReadTx and
creates a new one. This prevents onDataChanged from closing the Tx when
a read is in-flight.

Added a unit test class StatAbstractListenCommitTest with test cases for
the methods that were changed.

I also changed StatListenCommitFlow#initConfigFlows to create an
ArrayList when copying the Flow List as Linkedist incurs more memory
overhead.

Change-Id: I42e5b5d55eeddd62b0ffb002d35d07b664c91861
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>