Updated releng/autorelease Project: ovsdb  daa35f8eae02e128daa733e0ccf07182bbdd39a0
support for Interface external_ids

Adds support for Interface external_ids.  "external_ids" are configured on
Termination Point creation.  "external_ids" are persisted to the Operational
data store on Ovsdb Port update.  If external-id-value is not provided, the
port is still created as YANG has no way to enforce the field as mandatory.

Change-Id: Ibdac48f8d5144ddefd79278d5587c5a63bca0869
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>