Updated releng/autorelease Project: yangtools  bb3d28bb6172f1221ccdcb700c2f5d3d0d8de8d6
BUG-2955: Fix error reporting for wrong YangInstanceIdentifier

As turns out, attempting to load the wrong value will be reported via
UncheckedExecutionException, which we failed to account for. Make sure
we also catch this type and convert it into an Optional.absent() -- that
will trigger callers to properly report an IllegalArgumentException.

Change-Id: I50441cce4be94a2ef0fcf12591c6ec0e437f1084
Signed-off-by: Robert Varga <rovarga@cisco.com>